Rename `rasterizer.h` to `srp.h`
